Add documentation

+++ LilyPadEngine/src/LilyPad/debug/fps.hpp
@@ -7,17 +7,22 @@ namespace LilyPad
 	class FPS
 	{
 	public:
-
 		FPS();
 
 		// Update the frame rate
 		void update();
 
-		float get_delta();
+		/**
+		 * @brief Gets the delta between frames.
+		 *
+		 * @return double Time between frames in terms of seconds.
+		 */
+		double get_delta();
+
 	private:
 		// Time points to manage timing
-		float lastFrameTime;
-		float currentFrameTime;
-		float frameRate;
+		double lastFrameTime;
+		double currentFrameTime;
+		double frameRate;
 	};
 } // namespace LilyPad
